added material dialog library. field name can be changed directly by clicking on a label.

+++ app/src/main/kotlin/kr/ac/snu/hcil/omnitrack/core/database/DatabaseHelper.kt
@@ -383,9 +383,10 @@ class DatabaseHelper(context: Context) : SQLiteOpenHelper(context, "omnitrack.db
 
         val removedTrackerIds = user.fetchRemovedTrackerIds()
 
-        writableDatabase.delete(AttributeScheme.tableName, removedTrackerIds.map { "${AttributeScheme.TRACKER_ID} = ${it.toString()}" }.joinToString(" OR "), null)
-
-        deleteObjects(TrackerScheme, *removedTrackerIds)
+        if (!removedTrackerIds.isEmpty()) {
+            writableDatabase.delete(AttributeScheme.tableName, removedTrackerIds.map { "${AttributeScheme.TRACKER_ID} = ${it.toString()}" }.joinToString(" OR "), null)
+            deleteObjects(TrackerScheme, *removedTrackerIds)
+        }
 
         for (child in user.trackers.iterator().withIndex()) {
             save(child.value, child.index)
